+++ openredalert/src/game/CellTrigger.h
@@ -20,15 +20,17 @@
 
 #include <string>
 
+#include "SDL/SDL_types.h"
+
 using std::string;
 
 class CellTrigger
 {
 public:
     string name;
-    unsigned int cellpos;
-    unsigned int x;
-    unsigned int y;
+    Uint16 cellpos;
+    Uint16 x;
+    Uint16 y;
 };
 
 #endif //CELLTRIGGER_H
